Add base table directive with primitive CSV export.
The idea here was to create a proof of concept that would suffice to showcase the markup and controller logic.  This was a very haphazard process as I'm still figuring out AngularJS directives.  I have quite a bit of cleanup left to do in the angular-table.js module, even though this isn't a huge amount of code.